+++ ghost/api-framework/lib/frame.js
