+++ ghost/api-framework/lib/headers.js
