+++ ghost/api-framework/lib/pipeline.js
