+++ ghost/api-framework/lib/serializers/index.js
