+++ ghost/api-framework/lib/serializers/output/index.js
