+++ ghost/api-framework/lib/utils/options.js
