+++ ghost/api-framework/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@tryghost/api-framework",
+  "version": "0.0.0",
+  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/api-framework",
+  "author": "Ghost Foundation",
+  "private": true,
+  "main": "index.js",
+  "scripts": {
+    "dev": "echo \"Implement me!\"",
+    "test": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura mocha './test/**/*.test.js'",
+    "lint:code": "eslint *.js lib/ --ext .js --cache",
+    "lint": "yarn lint:code && yarn lint:test",
+    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
+  },
+  "files": [
+    "index.js",
+    "lib"
+  ],
+  "devDependencies": {
+    "c8": "7.12.0",
+    "mocha": "10.0.0",
+    "should": "13.2.3",
+    "sinon": "14.0.0"
+  },
+  "dependencies": {
+    "@tryghost/debug": "0.1.18",
+    "@tryghost/errors": "1.2.15",
+    "@tryghost/promise": "0.1.21",
+    "@tryghost/tpl": "0.1.18",
+    "@tryghost/validator": "0.1.27",
+    "jsonpath": "1.1.1",
+    "lodash": "4.17.21"
+  }
+}
