+++ ghost/core/.c8rc.json
@@ -8,7 +8,7 @@
   ],
   "statements": 57,
   "branches": 85,
-  "functions": 52,
+  "functions": 51,
   "lines": 57,
   "include": [
     "core/{*.js,frontend,server,shared}"
