+++ ghost/core/core/server/api/index.js
@@ -1,3 +1,1 @@
 module.exports.endpoints = require('./endpoints');
-
-module.exports.shared = require('./shared');
