+++ ghost/core/core/server/services/webhooks/serialize.js
@@ -2,7 +2,7 @@ module.exports = (event, model) => {
     const _ = require('lodash');
     const {sequence} = require('@tryghost/promise');
     const api = require('../../api').endpoints;
-    const apiShared = require('../../api').shared;
+    const apiFramework = require('@tryghost/api-framework');
 
     const resourceName = event.match(/(\w+)\./)[1];
     const docName = `${resourceName}s`;
@@ -23,7 +23,7 @@ module.exports = (event, model) => {
                 };
             }
 
-            return apiShared
+            return apiFramework
                 .serializers
                 .handle
                 .output(model, {docName: docName, method: 'read'}, api.serializers.output, frame)
@@ -46,7 +46,7 @@ module.exports = (event, model) => {
                 frame.options.withRelated = ['tags', 'authors'];
             }
 
-            return apiShared
+            return apiFramework
                 .serializers
                 .handle
                 .output(model, {docName: docName, method: 'read'}, api.serializers.output, frame)
