+++ ghost/core/core/server/web/api/endpoints/content/routes.js
@@ -1,7 +1,7 @@
 const express = require('../../../../../shared/express');
 const cors = require('cors');
 const api = require('../../../../api').endpoints;
-const http = require('../../../../api').shared.http;
+const {http} = require('@tryghost/api-framework');
 const mw = require('./middleware');
 const config = require('../../../../../shared/config');
