+++ ghost/core/package.json
@@ -55,6 +55,7 @@
     "@sentry/node": "7.9.0",
     "@tryghost/adapter-manager": "0.0.0",
     "@tryghost/admin-api-schema": "4.1.1",
+    "@tryghost/api-framework": "0.0.0",
     "@tryghost/api-version-compatibility-service": "0.0.0",
     "@tryghost/bookshelf-plugins": "0.5.0",
     "@tryghost/bootstrap-socket": "0.0.0",
